initial import of opfcli command line tool (#1)
* Initial import of opfcli command line tool

This is my attempt at implementing what we talked about in
operate-first/SRE#299. It's an integrated command line tool written in
Go that currently provides equivalents for:

- onboarding.sh (called create-project)
- enable-monitoring.sh
- operate-first/apps#626 (called grant-access)

This is my first time writing Go, ever, so there may be lots of
non-idiomatic code. In your reviews, be kind, but be thorough smile.

* Fix problems in pre-commit configuration

- Disable remove-tabs in pre-commit hook

  Standard go style includes using tabs for indentation. This is
  enforced by gofmt.

- Trailing whitespace in Markdown documents is significant

  We shouldn't be running the trailing-whitespace check on Markdown
  documents (because that's how you indicate a line break without
  introducing a new paragraph).