Refactor api into separate module (#10)
* Refactor api into a separate module

Make unit testing easier by separating the API from the command-line
processing logic. There is now an api module that defines an API type,
and API methods are called on objects of that type.

Cobra commands are allocated dynamically, rather than statically. If
we re-introduce tests specifically for the CLI, this will make testing
easier by ensuring that each test can start with a clean Cobra
environment.

Closes #9

* Add tests for api module

This is a separate commit so that it can be filtered out while
reviewing the code changes from the adjacent commits.

x-branch: feature/refactor-api