Add workflows for running pre-commit tests and building releases (#12)
* Add GitHub workflow to run pre-commit checks

This removes the prow-based CI machinery and replaces it with a
GitHub workflow (`.github/workflows/precommit.yml`) that runs on
pull requests and pushes.

The workflow runs a variety of linters as well as Go unit tests.

We take advantage of GitHub's caching support to improve the runtime
of the checks when the project dependencies haven't changed.

Closes #3

* Add GitHub workflow to generate release from tags

Add a workflow that runs on tags matching the pattern `v*` (e.g.,
`v0.0.1`). The workflow:

- Checks to make sure the tag is an annotated tag (because we use the
  tag description as the release description)
- Runs the pre-commit tests,
- Uses a matrix build to generate several binaries, and finally
- Uses [1] to create a release that includes the compiled binaries.

We take advantage of GitHub's caching support to improve the runtime
of the checks when the project dependencies haven't changed.

Closes #5

[1]: https://github.com/ncipollo/release-action

x-branch: feature/workflows